chore(product-analytics): move retention editor UI into the product folder
Moves retention's 12 remaining PA-owned files out of the shared
frontend/src/scenes/insights/ tree: eight pickers, the date picker and
the minimum-occurrences input into insights/retention/filters/, and the
two edit-sidebar panels into insights/retention/editor/.

MinimumOccurrencesInput is renamed RetentionMinimumOccurrencesInput; it
reads retentionLogic and has only ever had one consumer, so the neutral
name promised more than it delivered.

This completes retention as the first fully migrated insight workflow.
A retention-only change now reports 1 merge-queue target
(fe:product:product_analytics) instead of 92, because the lane is
computed from changed paths. The imports still point back at
insightLogic and insightVizDataLogic, so this buys queue parallelism,
not decoupling.

react-toastify moves into the product's peerDependencies, which the two
files that toast already needed.
