+++ frontend/src/lib/components/Cards/InsightCard/dashboardVisualizationOptions.tsx
@@ -5,7 +5,6 @@ import { Spinner } from '@posthog/lemon-ui'
 
 import { ChartFilter } from 'lib/components/ChartFilter/ChartFilter'
 import { LemonMenuItems } from 'lib/lemon-ui/LemonMenu'
-import { RetentionChartPicker } from 'scenes/insights/filters/RetentionChartPicker'
 import { insightLogic } from 'scenes/insights/insightLogic'
 import { insightVizDataLogic } from 'scenes/insights/insightVizDataLogic'
 
@@ -19,6 +18,8 @@ import {
     isTrendsQuery,
 } from '~/queries/utils'
 import { ChartDisplayType, InsightType, RetentionDashboardDisplayType } from '~/types'
+
+import { RetentionChartPicker } from 'products/product_analytics/frontend/insights/retention/filters/RetentionChartPicker'
 
 import { DashboardSqlChartType, DashboardSqlDisplayOptions } from './DashboardSqlDisplayOptions'
