+++ frontend/src/queries/nodes/InsightViz/InsightDisplayConfig.tsx
@@ -11,10 +11,8 @@ import { NON_TIME_SERIES_DISPLAY_TYPES } from 'lib/constants'
 import { LemonDropdown } from 'lib/lemon-ui/LemonDropdown'
 import { alignResolvedDateRangeToInterval, formatResolvedDateRange } from 'lib/utils/datetime'
 import { InsightDateFilter } from 'scenes/insights/filters/InsightDateFilter'
-import { RetentionChartPicker } from 'scenes/insights/filters/RetentionChartPicker'
 import { insightLogic } from 'scenes/insights/insightLogic'
 import { insightVizDataLogic } from 'scenes/insights/insightVizDataLogic'
-import { RetentionDatePicker } from 'scenes/insights/RetentionDatePicker'
 import { PathStepPicker } from 'scenes/insights/views/Paths/PathStepPicker'
 
 import { hasBreakdownFilter, isWebAnalyticsInsightQuery } from '~/queries/utils'
@@ -24,6 +22,8 @@ import { FunnelBinsPicker } from 'products/product_analytics/frontend/insights/f
 import { FunnelDisplayLayoutPicker } from 'products/product_analytics/frontend/insights/funnels/filters/FunnelDisplayLayoutPicker'
 import { funnelDataLogic } from 'products/product_analytics/frontend/insights/funnels/funnelDataLogic'
 import { RetentionBreakdownFilter } from 'products/product_analytics/frontend/insights/retention/filters/RetentionBreakdownFilter'
+import { RetentionChartPicker } from 'products/product_analytics/frontend/insights/retention/filters/RetentionChartPicker'
+import { RetentionDatePicker } from 'products/product_analytics/frontend/insights/retention/filters/RetentionDatePicker'
 
 import { useInsightDisplayOptions } from './insightDisplayOptions'
 import { InsightDisplayOptionsPanel } from './InsightDisplayOptionsPanel'
