+++ products/product_analytics/frontend/insights/retention/filters/RetentionDatePicker.tsx
@@ -6,7 +6,7 @@ import { Tooltip } from '@posthog/lemon-ui'
 import { DateFilter } from 'lib/components/DateFilter/DateFilter'
 import { insightLogic } from 'scenes/insights/insightLogic'
 
-import { retentionLogic } from 'products/product_analytics/frontend/insights/retention/retentionLogic'
+import { retentionLogic } from '../retentionLogic'
 
 export function RetentionDatePicker(): JSX.Element {
     const { insightProps, editingDisabledReason } = useValues(insightLogic)
